+++ src/index.test.ts
@@ -18,7 +18,7 @@ class ShortCircuitPlugin implements RateLimiterPlugin {
   }
 
   async hash() {
-    return this.value ?? false;
+    return this.value;
   }
 }
 
@@ -214,7 +214,7 @@ describe('Basic rate limiter', async () => {
       event = mockEvent() as RequestEvent;
     });
 
-    it.only('should always allow the request when true is returned and the plugin is first in the chain', async () => {
+    it('should always allow the request when true is returned and the plugin is first in the chain', async () => {
       const limiter = new RateLimiter({
         plugins: [new ShortCircuitPlugin(true, [1, 'm'])],
         rates: {
@@ -227,7 +227,7 @@ describe('Basic rate limiter', async () => {
       expect(await limiter.isLimited(event)).toEqual(false);
     });
 
-    it.only('should always deny the request when false is returned and the plugin is first in the chain', async () => {
+    it('should always deny the request when false is returned and the plugin is first in the chain', async () => {
       const limiter = new RateLimiter({
         plugins: [new ShortCircuitPlugin(false, [1, 'm'])],
         rates: {
@@ -240,7 +240,7 @@ describe('Basic rate limiter', async () => {
       expect(await limiter.isLimited(event)).toEqual(true);
     });
 
-    it.only('should deny the request when it is returning false further down the chain, and the first plugin is ok', async () => {
+    it('should deny the request when it is returning false further down the chain, and the first plugin is ok', async () => {
       const limiter = new RateLimiter({
         plugins: [new ShortCircuitPlugin(false, [5, 'm'])],
         rates: {
@@ -253,7 +253,7 @@ describe('Basic rate limiter', async () => {
       expect(await limiter.isLimited(event)).toEqual(true);
     });
 
-    it.only('should allow the request when it is returning true further down the chain, until the first plugin is limiting', async () => {
+    it('should allow the request when it is returning true further down the chain, until the first plugin is limiting', async () => {
       const limiter = new RateLimiter({
         plugins: [new ShortCircuitPlugin(true, [5, 'm'])],
         rates: {
@@ -261,6 +261,22 @@ describe('Basic rate limiter', async () => {
         }
       });
 
+      expect(await limiter.isLimited(event)).toEqual(false);
+      expect(await limiter.isLimited(event)).toEqual(false);
+      expect(await limiter.isLimited(event)).toEqual(true);
+    });
+
+    it('should deny the request when it is returning null further down the chain, until any other plugin is limiting', async () => {
+      const limiter = new RateLimiter({
+        plugins: [new ShortCircuitPlugin(null, [3, 'm'])],
+        rates: {
+          IP: [5, 'm']
+        }
+      });
+
+      expect(await limiter.isLimited(event)).toEqual(false);
+      expect(await limiter.isLimited(event)).toEqual(false);
+      expect(await limiter.isLimited(event)).toEqual(false);
       expect(await limiter.isLimited(event)).toEqual(false);
       expect(await limiter.isLimited(event)).toEqual(false);
       expect(await limiter.isLimited(event)).toEqual(true);