+++ CHANGELOG.md
@@ -5,6 +5,20 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.2] - 2023-07-11
+
+### Changed
+
+- Removed `check` method from `RateLimiterStore` interface.
+
+### Added
+
+- `RateLimiterPlugin` can now return `null`, as an indeterminate result.
+
+### Fixed
+
+- `RateLimiter` plugin chain wasn't immutable.
+
 ## [0.3.1] - 2023-07-02
 
 ### Security